Add optional automatic renaming to intro.
This introduce the possibility to add the flag ? after the ident when you introduce
something, if you don't it behaves as before.

If you put a ? and the ident already exists it will add a number at the end.

For example intros x? x? x? x?. Will introduce with idents x x0 x1 x2.