Modifications in intros : Replace ? by ! for introducing fresh names, and use ? for finding automatically the name.
If you do intros ? ? ?. and it will find the name in your goal for you(H for hyp)
and add a number if needed.
If you do intros x y z. It will work as before, introducing only if the name
isn't already used.
If you do intros x! y! z!. It will introduce vars even if x y z is already present
but will happen a number to it.

It works in all intros patterns.