Addition of basic morphism (without morphism lemmas) between FSets and finite ISets.
(==) is used to denote extensional equality of an ISet (on the left)
and an FSet (on the right). The notation does not make it clear that
the predicate has no symmetry, so a better one might be needed.

finite is an existentially quantified version of (==) (i.e. we define
an ISet as finite iff it is extensionally equal to some FSet).

toFSet takes an ISet and produces an FSet. If its argument is finite,
then the argument and result are extensionally equal.