Rewriting the PSS proof to introduce the looped sampling operation at the end rather than early on.
This should make it easier to understand where the proof is going in its early stages,
although it may make it more difficult to prove the loop transition (which is already harsh).
One possible intermediate game would be to sample from the uniform distribution on the preimage
by (finv sk) of the signatures that start with a 0 bit, before moving on to the actual upto bad
step, where the failure event is to sample u outside of that preimage.