Restructuring doc and some writing.
New global structure is now:
1. getting started
2. writing specifications
3. writing proofs
4. structuring specifications and proofs
5. examples and other stuff

Inside "writing specifications", I have also moved any mentions to theorems (other than illustrative example) to the end so as to have a more comprehensive description of the interactions between the different logics.
Rewriting and restructuring 3 is next.
Writing contents for Chapter 4 is now waiting on Pierre-Yves's implementation (and related discussions).