Named unary and binary op.
Named operators are of the form \xxx where xxx is any valid identifier.
As for other uniary/binary operators, they must be resp. enclosed in
(...) and [...]. E.g.

{{{
op (\in) (x : 'a) (xs : 'a list).
lemma L: forall (x : 'a), x \in [x].
}}}

They are left associative and have the highest priority among binary
operators.