New SMT tactic.
This commit brings a new `smt` tactic that can be called via `smt:lazy`.

This tactic relies on a lazy translation of operators/lemmas and is
currently missing:

  - a proper selection of lemmas to sent
  - the translation of body of fixpoints
  - a set of flags to better control the translation (including the
    use of the already existing hints databases)

The default `smt` version to be called can be controlled via
the `Smt:lazy` option. By default, the legacy version is used. If
changed to the new one, the old one can be locally called via
the `smt:full` variant of `smt`.

This commit also bring a `time` meta command that can be used
to time any vernacular command by simply executing:

  `time thecommand.`