Pretty-printing of notations.
The algorithm is quite brutal as it matches the notations' body
one by one, in reverse order of their definition. A better selection
algorithm can be implemented in case the pretty-printer starts to
be too slow in real devs using numerous notations.

The parsing of notations is still not implemented :)