New tactic: solve
`solve` is a resolution tactics. Currently, it tries to (recursively)
close goals with the lemmas in the hint databases (in the same manner
as `done` with the `hint exact` database).

However, `solve` can be given hint databases names with the syntax:

  `solve (db1, db2, ...)`

The maximum depth of the resolution tree is controled by an optional
argument that comes first (`solve 2` for example). By default, the
depth is equal to 1.

Databases can be created using the 'hint solve' command:

    `hint solve $priority $dbname : lemma1 lemma2'.

where the priority is optional --- `solve` tries to solve goals
using lemmas with a lower $priority first.

`hint exact ...` is now an alias for `hint solve 0 ...`