Add a simple form of pattern selection in rewrite rules
Syntax is: `rewrite [pattern]rule`

In this form, `pattern` is first searched against the
current goal and rule LHS is then matched against the
instanciated pattern. E.g.

  `rewrite [y+_]addrC`

will instantiate `addrC` to `y` and `z` assuming that
the first match of `y+_` in the current goal is `y+z`.

This syntax is compatible with other variants:

  `rewrite -{2}[y+_]addrC`