Extend `rnd` tactic s.t. it can handle multiple samplings / assignments
Syntax is `rnd ... : pos1 pos2` and is only evailable for the pRHL
variant.

The tactic will first collapse all the instructions after `pos1` /
`pos2` (in the left/right programs) before applying the `rnd`
tactic.

It is also possible to access this collapse phase with the new
`rndsem` tactic. The syntax is: `rndsem side? codepos`