Read library paths from the EC_IDIRS/EC_RDIRS env. vars.
Library paths can be specified using the environment variables EC_IDIRS
and EC_RDIRS. They contain each a list of specifications separated by
semicolumns (;).

Each said specification is a path, optionally qualified (when present,
the qualifier is followed by a colon (:)).

Specifications given in the EC_IDIRS variable behave the same as the
ones given in the idirs configuration entry or on the command line
following a -I argument.

Specifications given in the EC_RDIRS variable behave the same as the
ones given in the rdirs configuration entry or on the command line
following a -R argument.