Support project-local ini files
EasyCrypt now reads project-local files for getting its configuration
settings. This configuration file takes precedence over the
system-level configuration file.

The project-local configuration file is named 'easycrypt.project'.  It
is identical to the system-level configuration file.

When compiling a file, the configuration file is searched upward from
the target file directory. When ran in interactive mode, the
configuration file is searched upward from the current working
directory.

Note that when ProofGeneral starts EasyCrypt, the working directory is
set to the directory of the focused file.