Add tactic `weakmem`
This tactic allows to weaken the memory of a phoare hypothesis
by adding new variables to it.

This tactic is needed in the while rule for phoare, when one
wants to apply the induction hypothesis & the memories are
not compatible. Currently, the convertibility check does not
enforce that memories are equal, but a to-come PR is going to
fix that point.