New command prefix: `fail`
This prefix indicates that the following command should fail and that
the error has to be ignored. It is an actual error if the command
does not fail.

In interactive mode, the error is printed (at [info] level).

The main purpose of `fail` is to allow to write more precise unit
tests where one can indicate which command shoud actually fail.