internal: handle CTRL+C properly w.r.t. Why3 server
For that purpose, we now start Why3 manually and connect to it
as if it were an external Why3 server instance.

The Why3 server is put in its own process groups and does not receive
the signals (SIGINT, SIGTERM e.g.) directly received by EasyCrypt.

The Why3 server is started s.t. it shutdowns itself gracefully
when EasyCrypt disconnects from it.