Fix logical variable shadowing program variables
Currently, when resolving an identifier, EasyCrypt tries to resolve it
as a logical variable first, and then as a program variable (in the
active memory), even when a memory specifier is given.

This commit adds a new memory specifier that changes the resolution
order: when using it, we first try to resolve the identifier as a
program variable (in the given memory), and then as a logical variable
if no program variable matching that name is found.

The syntax if `form{!memory}`.

The pretty-printer has been updated accordingly.