Merge typing of expressions and formulas
This is a step forward removing expressions from EasyCrypt.

This commit removes all parsing and typing code specific to
expressions. Expressions are now parsed as formulas, utilizing the
formula type checker instead of an expression-specific type
checker. If a formula does not represent a valid expression, it is the
type checker's responsibility to raise a meaningful error message.