Fix bug in SMT translation (abstraction of non-translatable constructions)
When abstracting out a construction that is not translatable into SMT,
make the abstraction symbol depends on the local variables.

If not, the translation is unsound by making the translated constructions
a constant. For example, the following (where G will be abtracted):

```
fun x => G x
```

is translated into

```
fun x => fresh-constant
```