build docker containers in CI
We now build `base`, `build` and `formosa` containers in CI, and push
them to GHCR. This is done for the `main` and `release` branches, as
well as for tagged releases (`ryyyy.mm`). Containers are tagged with the
originating branch or tag when pushed.

This ensures that the software environment for building EasyCrypt (and
Jasmin, in the `formosa` container) is captured at time of release.